DLT SomeIP: Gather Error parsing in shared module
Set used Error parsing functions in one module to be shared in
indexer CLI too.
