+++ application/apps/indexer/indexer_cli/src/interactive.rs
@@ -1,9 +1,12 @@
 use crate::{duration_report, Instant};
 use futures::{pin_mut, stream::StreamExt};
-use parsers::{dlt::DltParser, LogMessage, MessageStreamItem, ParseYield};
+use parsers::{dlt::DltParser, MessageStreamItem, ParseYield};
 use processor::grabber::LineRange;
 use rustyline::{error::ReadlineError, DefaultEditor};
-use session::session::Session;
+use session::{
+    parse_err::{get_log_text, ParseErrorReslover},
+    session::Session,
+};
 use sources::{
     factory::{DltParserSettings, FileFormat, ObserveOptions, ParserType},
     producer::MessageProducer,
@@ -46,6 +49,7 @@ pub(crate) async fn handle_interactive_session(input: Option<PathBuf>) {
                             let udp_source = UdpSource::new(RECEIVER, vec![]).await.unwrap();
                             let dlt_parser = DltParser::new(None, None, None, false);
                             let mut dlt_msg_producer = MessageProducer::new(dlt_parser, udp_source, None);
+                            let mut parse_reslover = ParseErrorReslover::new();
                             let msg_stream = dlt_msg_producer.as_stream();
                             pin_mut!(msg_stream);
                             loop {
@@ -57,11 +61,11 @@ pub(crate) async fn handle_interactive_session(input: Option<PathBuf>) {
                                     item = msg_stream.next() => {
                                         match item {
                                             Some((_, MessageStreamItem::Item(ParseYield::Message(item)))) => {
-                                                let msg = get_msg_todo(item);
+                                                let msg = get_log_text(item, &mut parse_reslover);
                                                 println!("msg: {msg}");
                                             }
                                             Some((_, MessageStreamItem::Item(ParseYield::MessageAndAttachment((item, attachment))))) => {
-                                                let msg = get_msg_todo(item);
+                                                let msg = get_log_text(item, &mut parse_reslover);
                                                 println!("msg: {msg}, attachment: {attachment:?}");
                                             }
                                             Some((_, MessageStreamItem::Item(ParseYield::Attachment(attachment)))) => {
@@ -195,23 +199,4 @@ async fn collect_user_input(tx: mpsc::UnboundedSender<Command>) -> JoinHandle<()
         }
         println!("done with readline loop");
     })
-}
-
-//TODO AAZ: This should handle the cases when we have nested parsers + Saving the messages into
-///the faulty messages stack if the nested parsers couldn't resolve it.
-pub fn get_msg_todo(item: impl LogMessage) -> String {
-    let text_res = item.to_text();
-    if item.can_error() {
-        //TODO AAZ: Manage someip parser case for now
-        let mut msg = text_res.msg;
-        if let Some(err_info) = text_res.error {
-            msg = format!(
-                "{msg}: TODO: These bytes should be parsed with someip {:?}",
-                err_info.remain_bytes
-            );
-        }
-        msg
-    } else {
-        text_res.msg
-    }
 }
