+++ application/apps/indexer/session/src/handlers/mod.rs
@@ -2,7 +2,6 @@ pub mod export_raw;
 pub mod extract;
 pub mod observe;
 mod observing;
-pub mod parse_err;
 pub mod search;
 pub mod search_values;
 pub mod sleep;
