+++ application/apps/indexer/session/src/parse_err.rs
@@ -40,3 +40,27 @@ impl ParseErrorReslover {
         }
     }
 }
+
+/// Get the text message of [`LogMessage`], resolving parse text errors if possible,
+/// TODO: Otherwise it should save the error to the faulty messages store, which need to be
+/// implemented as well :)
+pub fn get_log_text(item: impl LogMessage, err_resolver: &mut ParseErrorReslover) -> String {
+    let text_res = item.to_text();
+    if item.can_error() {
+        let mut msg = text_res.msg;
+        if let Some(err_info) = text_res.error {
+            match err_resolver.resolve_err(&err_info) {
+                Some(resloved_msg) => {
+                    msg.push_str(&resloved_msg);
+                }
+                None => {
+                    //TODO: Item with error details should be reported faulty messages store.
+                    msg = format!("{msg}: Unknow Error bytes: {:?}", err_info.remain_bytes);
+                }
+            }
+        }
+        msg
+    } else {
+        text_res.msg
+    }
+}
